temp

+++ lib/readData.h
@@ -11,7 +11,7 @@
 namespace readData {
     class readData {
     public:
-        static std::vector<std::vector<double> > readImageData(std::string path, int size = -1);
+        static std::vector<std::vector<float> > readImageData(std::string path, int size = -1);
 
         static std::vector<int> readTagData(std::string path,int size = -1);
     };
